Update script that prepares mongo db for use. Use existing configuration if one exists.

+++ sbin/configure_mongo_dev.sh
@@ -15,44 +15,62 @@ if [ ! -f $MONGO_BIN_DIRECTORY/mongosh ]; then
   exit 1
 fi
 
-# Check to make sure there is nothing in the mongo directory yet
-if [ "$(ls -A "$MONGO_DATA_DIRECTORY")" ]; then
-  >&2 echo "Aborting: MongoDB has data in the data directory: $MONGO_DATA_DIRECTORY"
-  exit 1
-fi
+if [[ -f $MONGO_ADMIN_PASSWD_FILE && -f $MONGO_TRAVELER_PASSWD_FILE ]]; then
+  # --- Reconfigure mode: existing credentials found, prepare a clean database ---
+  echo "Existing configuration found. Preparing a clean database..."
 
-read -p "What port would you like mongodb to run on? [27017]: " mongoPort
-if [[ -z $mongoPort ]]; then
-  mongoPort=27017
-fi
-if [ ! -z $mongoPort ]; then
-  mongoConfigurationFile=$TRAVELER_INSTALL_ETC_DIR/mongo-configuration.sh
-  cmd="cat $mongoConfigurationFile | sed 's?export MONGO_SERVER_PORT=.*?export MONGO_SERVER_PORT=$mongoPort?g' > $mongoConfigurationFile.2 && mv $mongoConfigurationFile.2 $mongoConfigurationFile"
-  eval $cmd
-  export MONGO_SERVER_PORT=$mongoPort
+  adminPass=`cat $MONGO_ADMIN_PASSWD_FILE`
+  travelerPass=`cat $MONGO_TRAVELER_PASSWD_FILE`
+
+  $TRAVELER_ETC_INIT_DIRECTORY/traveler-mongodb stop || true
+
+  if [ "$(ls -A "$MONGO_DATA_DIRECTORY")" ]; then
+    >&2 echo "Error: $MONGO_DATA_DIRECTORY is not empty."
+    >&2 echo "Please manually remove its contents before re-running this script."
+    exit 1
+  fi
+else
+  # --- Fresh setup mode ---
+
+  # Check to make sure there is nothing in the mongo directory yet
+  if [ "$(ls -A "$MONGO_DATA_DIRECTORY")" ]; then
+    >&2 echo "Aborting: MongoDB has data in the data directory: $MONGO_DATA_DIRECTORY"
+    exit 1
+  fi
+
+  read -p "What port would you like mongodb to run on? [27017]: " mongoPort
+  if [[ -z $mongoPort ]]; then
+    mongoPort=27017
+  fi
+  if [ ! -z $mongoPort ]; then
+    mongoConfigurationFile=$TRAVELER_INSTALL_ETC_DIR/mongo-configuration.sh
+    cmd="cat $mongoConfigurationFile | sed 's?export MONGO_SERVER_PORT=.*?export MONGO_SERVER_PORT=$mongoPort?g' > $mongoConfigurationFile.2 && mv $mongoConfigurationFile.2 $mongoConfigurationFile"
+    eval $cmd
+    export MONGO_SERVER_PORT=$mongoPort
+  fi
+
+  # Get password for the users that will be created
+  echo "Admin password will be stored in a passwd file: $MONGO_ADMIN_PASSWD_FILE"
+  read -s -p "Enter admin password for mongodb(it will be stored in a config file) [admin]: " adminPass
+  echo ''
+  read -s -p "Enter traveler db password for mongodb(it will be stored in a config file) [traveler]: " travelerPass
+  echo ''
+
+  if [[ -z $adminPass ]]; then
+    adminPass="admin"
+  fi
+  if [[ -z $travelerPass ]]; then
+    travelerPass="traveler"
+  fi
+
+  echo $adminPass > $MONGO_ADMIN_PASSWD_FILE
+  chmod 400 $MONGO_ADMIN_PASSWD_FILE
+  echo $travelerPass > $MONGO_TRAVELER_PASSWD_FILE
+  chmod 400 $MONGO_TRAVELER_PASSWD_FILE
 fi
 
 # Start mongodb server
 $TRAVELER_ETC_INIT_DIRECTORY/traveler-mongodb startNoAuth
-
-# Get password for the users that will be created
-echo "Admin password will be stored in a passwd file: $MONGO_ADMIN_PASSWD_FILE"
-read -s -p "Enter admin password for mongodb(it will be stored in a config file) [admin]: " adminPass
-echo ''
-read -s -p "Enter traveler db password for mongodb(it will be stored in a config file) [traveler]: " travelerPass
-echo ''
-
-if [[ -z $adminPass ]]; then
-  adminPass="admin"
-fi
-if [[ -z $travelerPass ]]; then
-  travelerPass="traveler"
-fi
-
-echo $adminPass > $MONGO_ADMIN_PASSWD_FILE
-chmod 400 $MONGO_ADMIN_PASSWD_FILE
-echo $travelerPass > $MONGO_TRAVELER_PASSWD_FILE
-chmod 400 $MONGO_TRAVELER_PASSWD_FILE
 
 commandjs="use admin;"
 commandjs="$commandjs \n db.createUser({ user: \"$MONGO_ADMIN_USERNAME\", pwd: \"$adminPass\", roles: [ \"root\" ] } )"
